Changes to remove anti pattern
